Add a CompiledRequirements model
The goal is to save both the results of pip-compile and (when
implemented), of our custom optimized version. This would allow in a
first phase to always return the pip-compile version and gather feedback
about effectiveness and speed of our custom solution when compared to
the original version.

We can also rerun the optimized version compilation and test it against
the original version.

In a second phase, it would be possible to gradually switch from one
version to the other by simply switching which version is returned to
the client.

In a last phase, we would remove the model and always return the
optimized version on the fly.