Correctly import cached packages
The new PyPI infrastructure heavily caches package info endpoints.
It can occur that we receive a notification for an updated package and
we get the package info before its cache is busted, effectively syncing
our local state with the old cached version.

This change ensures that the returned info is up to date with the event
serial which first introduced the change.